cleanup: 70 clippy warnings to 12, and four invariants now fail at compile time
The five "assertion has a constant value" warnings were the interesting ones.
Every term in them is a constant — MIN_BUTTON_HOLD against Apple's 150 ms diag
poll, and the four that hold the iPod's hold switch on the top edge and inside
the case — so the assertions could only ever fail when somebody ran the suite.
Moved into `const` blocks, they fail the BUILD. An invariant over constants that
waits for a test run is weaker than it looks, and this project has been bitten
before by a check that was present, correct in isolation, and never reached.

Also: a stray second `use super::*;` sitting in the middle of emu.rs's test
module, and `cargo clippy --fix` over the mechanical rest — useless vec!, manual
is_multiple_of, redundant closures, identity maps, needless returns. settings.rs
and main.rs show large diffs that are pure moves: `repo_root` and friends were
declared after the test module.

299 tests pass, as before.

+++ tools/eapp-loader/src/inspect.rs
@@ -980,7 +980,7 @@ mod naming_tests {
 
     /// Write a `SysCfg` block into a NOR image, in the layout [`syscfg`] reads.
     fn write_syscfg(nor: &mut [u8], serial: Option<&str>, guid: Option<u64>) {
-        let block = &mut nor[SYSCFG_AT as usize..];
+        let block = &mut nor[SYSCFG_AT..];
         block[..4].copy_from_slice(SYSCFG_MAGIC);
         let mut count = 0u32;
         let mut at = SYSCFG_HEADER;
